用户模块完成，转为传感器模块开发

+++ java_system/src/main/java/com/xm/mapper/UserMapper.java
@@ -14,5 +14,5 @@ public interface UserMapper {
     User findUserByPassword(@Param("id") long id,@Param("gId") int gId);
     List<User> getUserList(@Param("gId") int gId );
     Boolean addUser(@Param("user") User user);
-    Boolean deleteUserById (@Param("userId") int userId ,@Param("gId") int gId);
+    Boolean deleteUserById (@Param("userId") long userId ,@Param("gId") int gId);
 }
